Move declaration and statement module into model package
Ref. #208
